LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-area-computed-insets-expected.txt


FAIL position-area does not affect insets at computed value time assert_equals: position-area is supported expected "span-all" but got "none"
